Activate venvs for relevant commands
Since each command in a recipe runs in a separate shell, activating a
venv has no effect on subsequent commands. This fixes that by running
venv-relative commands in the same recipe "line".

While we're at it, make docs the default target, and ensure that the
requirements are installed before attempting to run mkdocs.

Signed-off-by: Stephen Kitt <skitt@redhat.com>